push assertContentType up to Api/Test case, remove DocData test case. If it's necessary in the future, recreate it.
